backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
slowapi==0.1.9
aiofiles==23.2.1
uuid6==1.1.1
